requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pymysql>=1.1
flask>=2.3
pyahocorasick>=2.0
lxml>=4.9